modify for JDK6
git-svn-id: http://192.168.20.11/sequoiadb/trunk@32563 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ driver/java/src/test/java/com/sequoiadb/test/cl/DBCollectionTest.java
@@ -473,7 +473,7 @@ public class DBCollectionTest {
         popOption.put("LogicalID", 1024);
         clCapped.pop(popOption);
         BSONObject record = clCapped.queryOne();
-        assertEquals(2048, (long)( record.get("_id")));
+        assertEquals(2048, ((Long)( record.get("_id"))).longValue());
 
         BSONObject popOption2 = new BasicBSONObject();
         popOption2.put("LogicalID", 2048);
